Improve handling of call trace size
- We read the size from env var GAUCHE_CALL_TRACE.
- Since error handling machinery makes quite a few calls internally,
which won't be shown in the trace becase they lack the source info,
too small number of CALL_TRACE doesn't show anything.  We need at least
several hundreds.  We set a minimum and maximum.
- Limit call trace size to 2^n.  Other values are rounded up to the closest
2^n in Scm__MakeCallTraceQueue.  Minimum and maximum are also enforced there.
So, the user can just set GAUCHE_CALL_TRACE=1 if they want to use minimum
value.

+++ src/gauche/priv/vmP.h
@@ -44,12 +44,12 @@ typedef struct ScmCallTraceEntryRec {
 } ScmCallTraceEntry;
 
 struct ScmCallTraceRec {
-    int size;                     /* size of the array */
-    int top;                      /* index to the next entry */
+    u_long size;                  /* size of the array, must be 2^n */
+    u_long top;                   /* index to the next entry */
     ScmCallTraceEntry entries[1]; /* variable length */
 };
 
-ScmCallTrace *Scm__MakeCallTraceQueue(int size);
+ScmCallTrace *Scm__MakeCallTraceQueue(u_long size);
 
 SCM_DECL_END
